+++ docs/tr1/CHANGELOG.md
@@ -8,6 +8,7 @@
     PS1     | 77  | 255   | 255  | ![#4DFFFF](https://placehold.co/15x15/4DFFFF/4DFFFF.png) `#4DFFFF`
     DOS     | 153 | 179   | 255  | ![#99B3FF](https://placehold.co/15x15/99B3FF/99B3FF.png) `#99B3FF`
 - changed the `draw_distance_min` and `draw_distance_max` to `fog_start` and `fog_end`
+- changed `Select Detail` dialog title to `Graphic Options`
 - fixed the bilinear filter to not readjust the UVs (#2258)
 - fixed anisotropy filter causing black lines on certain GPUs (#902)
 - fixed mesh faces not being drawn under some circumstances (#2452, #2438)